[#37] Refactor : 컴포넌트 기능별 모듈화 - 입력 - 결과 - 선택결과
